+++ api/src/main/java/ca/bc/gov/educ/api/batchgraduation/config/EducGradBatchApiConfig.java
@@ -1,0 +1,32 @@
+package ca.bc.gov.educ.api.batchgraduation.config;
+
+import net.javacrumbs.shedlock.core.LockProvider;
+import net.javacrumbs.shedlock.provider.jdbctemplate.JdbcTemplateLockProvider;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.transaction.PlatformTransactionManager;
+
+@Configuration
+public class EducGradBatchApiConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+        ModelMapper modelMapper = new ModelMapper();
+        return modelMapper;
+    }
+
+    /**
+     * Lock provider lock provider.
+     *
+     * @param jdbcTemplate       the jdbc template
+     * @param transactionManager the transaction manager
+     * @return the lock provider
+     */
+    @Bean
+    public LockProvider lockProvider(@Autowired JdbcTemplate jdbcTemplate, @Autowired PlatformTransactionManager transactionManager) {
+        return new JdbcTemplateLockProvider(jdbcTemplate, transactionManager, "BATCH_SHEDLOCK");
+    }
+}
